Also move out the exceptions
